Basic Level Calculator
This project encompasses a basic-level calculator developed in the Kotlin programming language. The calculator is capable of performing four fundamental mathematical operations (addition, subtraction, multiplication, division). With its user-friendly interface, it offers a simple and comprehensible user experience. The project adopts a modern approach, leveraging the advantages of the Kotlin language to provide reliable performance in calculations. In addition to executing basic mathematical operations, this project serves as a valuable resource for those looking to understand and learn the usage of the Kotlin language